ADD interpreter.rkt MOD test1.rkt
Just wanted to preface by saying I did a bunch of other stuff on the project for personal practice; feel free to disregard most of the contents in interpreter.rkt. Also the misspelling of the original file name was bugging me, so I deleted it and made a new file. Just working through tests and debugging as I go.