MOD interpreter.rkt MOD test1.txt
This is all of the state management mechanics updated to support a layered state space. Many of the original state management functions work as good helpers for modified the state by layer, so many of them have been changed to layer-<function-name>. See the comments above each function, it's a rather straightforward system. State is now saved as '(()) when 'intepret' is called and within the innermost parentheses, a new set of parenthesis containing a binding ex: (a 2) will be inserted when a binding is create, as so: '(((a 2))). If another binding were added in the same layer: '(((a 2) (b 2))). If a block were to be entered, and the block contained bindings to be created, it may look like this: '(((z 0)) ((a 2) (b 2))).